scripts/requirements.txt
pandas==1.5.3
requests==2.28.2
google-cloud-secret-manager==2.16.2